swval: purge gprof stuff from project..
gprof does not provide a means to collect profile data for .so libs,
only applications.
sprof can be used to collect profile data for .so's, but the output
format is not compatible with that of gprof. One cannot generate a big
picture data set that contains both the app and the .so data in one.
Forget gprof/sprof and use valgrind instead.