Progress on fixing the regression..
A NULL check for pProc as passed to OSLThread::OSLThread was recently
added. Seeing a NULL pProc, the constructor will exit early, prior to
setting m_IsOK to true. This introduced a bug with later manipulation of
m_Thread by the destructor, which only guarded m_Thread by
!m_LocalThread. The fix is to also guard m_Thread by m_IsOK, so that
pthreads API's are not called with an uninitialzed pthread_t.