Add pcipeek.py and aalinfo.
The pcipeek script replicates pcimem read functionality. It is limited to 64
bits. Usage:

   pcipeek.py <pci_resource> <offset>

The aalinfo script scans lspci output and looks for BDX-P or SKX-P physical
functions. If found, it reads the AFU_ID in PORT0 of that PF and prints it in
GUID notation. Offsets for AFU_ID are hardcoded, so this will likely break in
future instantiations when we have more than one AFU or the MMIO map of the
PORTs change.