Add filename prefix check to ALI_Configure_AFU
To help catch accidental use of incompatible green bitstreams, this adds
a check for the first 20 characters of a green bitstream file, which is

   090716_skxp_611_pr

for 6.1.1 green bitstreams (6.1.1 is the version of the BBS release, not
AAL!). This is obviously not a true safeguard, and it can be overriden
by giving a --force=TRUE command line option.

Remember to adjust for changing filenames (i.e. on a BBS version number
change).

Change-Id: I8f74d9b0e7d308ca403846ae5f6d07fddfd87199