Print error on local ResMgr allocation failure
Add a check to ServiceBroker to catch the case when we can't allocate a local
Resource Manager. If this is due to a failure to open the device to talk to the
remote resource manager, this usually means that the drivers are not loaded; in
this case, print an additional message.

Change-Id: I8136b9b08e4817e31789ee5c5b2b2bb36fc9b8de