Removes UnWrapTransactionIDFromEvent function
This function was originally identified in Klockwork issue 179 but 
upon inspection was found that this function was no longer used.
It was recommended to disable or remove this function from the code 
base along with any associated tests. 

This removes UnWrapTransactionIDFrom event function from the header
file and the implementation in AASLib. It also removes the
corresponding test from swval test suite.

Change-Id: Id475b61893be0f06784119c4c990ddbf1543b95e