undo delete of aal0811 and comment
I would prefer to comment out the test case.
This seems to be what Tim has done in the past.
It also preserves the code for my reference and
the case number progression.  I.e. aal0810
will not be followed by aal0812, but rather a
disabled aal0811 so it's clear what happened.

Change-Id: I33549b528a838db7386006b47b97aaaae615e6e9