Addresses KW 775 by removing DoRegistrar function
The body of this function is surrounded by #if 0...#endif and
didn't return a value, hence the KW warning about non-void
function not returning a value. The only time this function is
called is in a switch statement in the function Parse_AALRMS_Msg.
The call to the function was replaced with an error message
indicating that a depracated request code was used.

Change-Id: Ibe84ce82c15ac0e7188824773587233f203ee327