Fix memory leak in ALIAFUProxy.cpp:213
ALIAFUProxy::Release() calls UIDriverInterfaceAdapter::SendMessage()
passing a pointer to an UnBind message, the contents of which will be
copied out into an ioctl that is sent synchronously to the driver.
The contents of the UnBind message are not subsequently used, and so
it does not need to hang around.
The current code allocates it on the heap and never frees it.
The new code allocates it on the stack so that it will be freed.

Change-Id: I992039b17167a7d1d40ac8234fb0ba73569fbeaa