Fix memory leak in ALIAFUProxy.cpp:124
ALIAFUProxy::init() calls UIDriverInterfaceAdapter::SendMessage()
passing a pointer to a Bind message, the contents of which will be
copied out into an ioctl that is sent synchronously to the driver.
The contents of the Bind message are not subsequently used, and so
it does not need to hang around.
The current code allocates it on the heap and never frees it.
The new code allocates it on the stack so that it will be freed.

Change-Id: I155277851861bfa8862a1aa1ca96004c12b465ed