Check BBS interface ID against metadata
To prevent partial reconfiguration of green bitstreams that are not compatible
with the currently loaded blue bitstream, this change adds a check to
ALIConfAFU and ALI that checks an interface ID provided as part of the green
bitstream file against the FME_PR_INTFC_ID CSRs in the blue bitstream (FME). It
will abort reconfiguration if the interface IDs don't match.

Configuring incompatible bitstreams can have severe consequences, as they might
overwrite portions of the blue bitstream. Since with Alpha+ (6.3.0), customers
will gain the ability to generate their own green bitstreams, and we will
provide multiple blue bitstreams as part of the release, there is a risk that
incompatible bitstreams will be tried.

Because this change requires metadata added to the raw RBF bitstream file,
these augmented files are now called GBS files (with the respective .gbs
filename extension). The metadata format used for this change is an interim
solution for Alpha+ and is planned to be superseded by the actual packager
implementation and associated kernel-level code for metadata parsing. This
Alpha+ interim implementation is realized completely in user space (see caveats
below). It will reject attempts to load RBF files.

The interim metadata consists of the magic byte sequence 0x80 0x86 0x1f 0x1d,
followed by the 128 bit interface ID in little endian representation. It is
prepended directly to the raw RBF file to create a GBS file. This patch relies
on a corresponding change to the script used to generate green bitstreams.

This patch also replaces the previous filename prefix checks, which did not
provide adequate incomaptibility protection and introduced undesired
dependencies between blue bistream releases and AAL. As part of this, the
"--force" option has been removed from ALIConfAFU.

Change-Id: I0845b0cd2ed9d90ba26c22079054bf6ec5336505